internal refactor
