+++ lib/deferred.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var simulacra = require('./')
+var simulacra = require('./core')
 var processNodes = require('./process_nodes')
 var changeValue = require('./change_value')
 var keyMap = require('./key_map')
@@ -21,6 +21,13 @@ var rootMap = new WeakMap()
 
 // This maps data object to definition.
 var defMap = new WeakMap()
+
+// Option to use comment nodes as markers.
+Object.defineProperty(render, 'useCommentNode', {
+  get: function () { return processNodes.useCommentNode },
+  set: function (value) { processNodes.useCommentNode = value },
+  enumerable: true
+})
 
 
 module.exports = render
