+++ lib/index.js
@@ -1,194 +1,32 @@
 'use strict'
 
-var processNodes = require('./process_nodes')
+var simulacra = require('./core')
 var bindKeys = require('./bind_keys')
-var keyMap = require('./key_map')
-
-var isArray = Array.isArray
-var hasBindingsKey = keyMap.hasBindings
-var hasDefinitionKey = keyMap.hasDefinition
-var markerKey = keyMap.marker
-var replaceAttributeKey = keyMap.replaceAttribute
-var isBoundToParentKey = keyMap.isBoundToParent
-var isDeferredKey = keyMap.isDeferred
-
-// Node names which should have value replaced.
-var replaceValue = [ 'INPUT', 'TEXTAREA', 'PROGRESS' ]
-
-// Input types which use the "checked" attribute.
-var replaceChecked = [ 'checkbox', 'radio' ]
-
-var useCommentNode = false
-var retainElement = '__retain__' +
-  ('00000000' + Math.floor(Math.random() * Math.pow(2, 32)).toString(16))
-  .slice(-8)
-
-// Upgrade to Symbol if possible, it guarantees uniqueness.
-if (typeof Symbol === 'function') retainElement = Symbol(retainElement)
-
-// Symbol for retaining an element instead of removing it.
-bindKeys.retainElement = retainElement
-Object.defineProperty(simulacra, 'retainElement', {
-  enumerable: true, value: retainElement
-})
+var processNodes = require('./process_nodes')
 
 // Option to use comment nodes as markers.
-processNodes.useCommentNode = useCommentNode
-Object.defineProperty(simulacra, 'useCommentNode', {
-  get: function () { return useCommentNode },
-  set: function (value) {
-    processNodes.useCommentNode = useCommentNode = value
-  },
+Object.defineProperty(bindObject, 'useCommentNode', {
+  get: function () { return processNodes.useCommentNode },
+  set: function (value) { processNodes.useCommentNode = value },
   enumerable: true
 })
 
+bindKeys.retainElement = simulacra.retainElement
 
-// Entry point of the package.
-module.exports = simulacra
+
+// Default entry point of the package.
+module.exports = bindObject
 
 
 /**
- * Bind an object to the DOM.
- *
- * @param {Object} obj
- * @param {Object} def
- * @return {Node}
+ * Wrapper around internal function.
  */
-function simulacra (obj, def) {
-  var document = this ? this.document : window.document
-  var Node = this ? this.Node : window.Node
-  var node, query, path
+function bindObject (obj, def) {
+  var node = simulacra.call(this, obj, def)
+  var path = []
 
-  if (typeof obj !== 'object' || isArray(obj))
-    throw new TypeError('First argument must be a singular object.')
-
-  if (obj[hasBindingsKey])
-    throw new Error('Can not bind an object that is already bound.')
-
-  if (!isArray(def))
-    throw new TypeError('Second argument must be an array.')
-
-  if (typeof def[0] === 'string') {
-    query = def[0]
-    def[0] = document.querySelector(query)
-    if (!def[0]) throw new Error(
-      'Top-level node "' + query + '" could not be found in the document.')
-  }
-  else if (!(def[0] instanceof Node)) throw new TypeError(
-    'The first position of top-level must be a Node or a CSS selector string.')
-
-  ensureNodes(this, def[0], def[1])
-
-  node = processNodes(this, def[0].cloneNode(true), def[1])
-
-  if (!obj[isDeferredKey]) {
-    // Assign internal binding key to note that the object is already bound.
-    Object.defineProperty(obj, hasBindingsKey, { value: true })
-
-    // Initialize path and assign root object.
-    path = []
-    path.root = obj
-
-    bindKeys(this, obj, def[1], node, path)
-  }
+  path.root = obj
+  bindKeys(this, obj, def[1], node, path)
 
   return node
-}
-
-
-/**
- * Internal function to mutate string selectors into Nodes and validate that
- * they are allowed.
- *
- * @param {Object} [scope]
- * @param {Node} parentNode
- * @param {Object} def
- */
-function ensureNodes (scope, parentNode, def) {
-  var Element = scope ? scope.Element : window.Element
-  var adjacentNodes = []
-  var i, j, defKeys, key, query, branch, boundNode, ancestorNode
-
-  if (typeof def !== 'object') throw new TypeError(
-    'The second position must be an object.')
-
-  defKeys = Object.keys(def)
-
-  for (i = 0, j = defKeys.length; i < j; i++) {
-    key = defKeys[i]
-
-    if (!isArray(def[key]))
-      def[key] = [ def[key] ]
-
-    branch = def[key]
-
-    // Internal value used while processing nodes.
-    Object.defineProperty(branch, markerKey, { value: null, writable: true })
-
-    if (typeof branch[0] === 'string') {
-      query = branch[0]
-
-      // May need to get the node above the parent, in case of binding to
-      // the parent node.
-      ancestorNode = parentNode.parentNode || parentNode
-
-      branch[0] = ancestorNode.querySelector(query)
-      if (!branch[0]) throw new Error(
-        'The element for selector "' + query + '" was not found.')
-    }
-    else if (!(branch[0] instanceof Element))
-      throw new TypeError('The first position on key "' + key +
-        '" must be a DOM element or a CSS selector string.')
-
-    boundNode = branch[0]
-
-    if (typeof branch[1] === 'object' && branch[1] !== null) {
-      Object.defineProperty(branch, hasDefinitionKey, { value: true })
-      if (branch[2] && typeof branch[2] !== 'function')
-        throw new TypeError('The third position on key "' + key +
-          '" must be a function.')
-    }
-    else if (branch[1] && typeof branch[1] !== 'function')
-      throw new TypeError('The second position on key "' + key +
-        '" must be an object or a function.')
-
-    // Special case for binding to parent node.
-    if (parentNode === boundNode) {
-      Object.defineProperty(branch, isBoundToParentKey, { value: true })
-      if (branch[hasDefinitionKey]) ensureNodes(scope, boundNode, branch[1])
-      else if (typeof branch[1] !== 'function')
-        console.warn( // eslint-disable-line
-          'A change function was not defined on the key "' + key + '".')
-      continue
-    }
-    else adjacentNodes.push([ key, boundNode ])
-
-    if (!parentNode.contains(boundNode))
-      throw new Error('The bound DOM element must be either ' +
-        'contained in or equal to the element in its parent binding.')
-
-    if (branch[hasDefinitionKey]) {
-      ensureNodes(scope, boundNode, branch[1])
-      continue
-    }
-
-    Object.defineProperty(branch, replaceAttributeKey, {
-      value: ~replaceValue.indexOf(boundNode.nodeName) ?
-        ~replaceChecked.indexOf(boundNode.type) ?
-        'checked' : 'value' : 'textContent'
-    })
-  }
-
-  // Need to loop again to invalidate containment in adjacent nodes, after the
-  // adjacent nodes are found.
-  for (i = 0, j = defKeys.length; i < j; i++) {
-    key = defKeys[i]
-    boundNode = def[key][0]
-    for (i = 0, j = adjacentNodes.length; i < j; i++)
-      if (adjacentNodes[i][1].contains(boundNode) &&
-        adjacentNodes[i][1] !== boundNode)
-        throw new Error(
-          'The element for key "' + key + '" is contained in the ' +
-          'element for the adjacent key "' + adjacentNodes[i][0] + '".')
-  }
 }
