+++ lib/process_nodes.js
@@ -5,6 +5,9 @@ var keyMap = require('./key_map')
 var isBoundToParentKey = keyMap.isBoundToParent
 var markerKey = keyMap.marker
 
+// Option to use comment nodes as markers.
+processNodes.useCommentNode = false
+
 
 module.exports = processNodes
